lmer_uvcov {lme4GS}	R Documentation

Fits a linear mixed model with user specified variance covariance-matrices.

Description

Fits a linear mixed model with user specified variance covariance-matrices.

Usage

  lmer_uvcov(y,fixed="1",random,verbose=5L)

Arguments

y
(numeric, n) the data-vector (NAs not allowed).

fixed
A formula that specifies fixed effects.

random
(list) This is a two-level list used to specify the variance covariance matrixes for random effects.

verbose
verbose output from optimizeLmer function?.

Details

The routine fits the linear mixed model:
y=Xβ + Z1 u1 + ... + Zq uq + e,
where y is the response vector, X is the matrix for fixed effects, β is the vector of fixed effects, Zj is a design matrix for random effects, uj is a vector of random effects, j=1,...,q. We assume that uj~N(0,σ2j K j), j=1,...,q and e~N(0,σ2eI).

The linear mixed model can be re-written as:
y=y=Xβ + Z1* u1*+...+Zq* uq*+e,
where Zj*=Zj Lj, with Lj from Cholesky factorization for Kj. Alternatively, Zj*=ZjΓjΛ1/2, with Γj and Λj the matrix of eigen-vectors and eigen-values obtained from the eigen-value decomposition for Kj. The factorization method for Kj is selected automatically at runtime.

Value

An object of class merMod (more specifically, an object of subclass lmerMod), for which many methods are available (e.g. methods(class="merMod"))

Author(s)

Paulino Perez-Rodriguez

Examples

## Not run: 
  library(BGLR)
  library(lme4GS)

  #Example 1, wheat 
  data(wheat)
  X<-wheat.X
  Z<-scale(X,center=TRUE,scale=TRUE)
  G<-tcrossprod(Z)/ncol(Z)
  A<-wheat.A
  rownames(G)<-colnames(G)<-rownames(A)
  y<-wheat.Y[,1]

  #id a vector with ids
  random<-list(mrk=list(K=G,id=rownames(G)),
               ped=list(K=A,id=rownames(A)))

  out<-lmer_uvcov(y,fixed="1",random=random)

  summary(out)

  plot(y,predict(out))

  #Example 2, mice

  data(mice)

  y<-mice.pheno$Obesity.BMI
  y<-scale(y,center=TRUE,scale=TRUE)
  y<-as.vector(y)

  X<-mice.X
  Z<-scale(X,center=TRUE,scale=TRUE)
  G<-tcrossprod(Z)/ncol(Z)

  gender<-mice.pheno$GENDER
  litter<-mice.pheno$Litter
  cage<-mice.pheno$cage

  fixed<-"1+gender+litter"

  #K=NULL means that we are using the identity matrix
  random<-list(cage=list(K=NULL,id=cage),
	           mrk=list(K=G,id=rownames(G)))

  out<-lmer_uvcov(y,fixed=fixed,random=random)

  plot(y,predict(out))

## End(Not run)

[Package lme4GS version 1.0 Index]