maize.G {lme4GS}	R Documentation

Genomic relationship matrix for maize lines

Description

A matrix with relationship between individuals for parents of two heterotic groups. The matrix was computed using 511 SNPs using the function A.mat included in rrBLUP package (Endelman, 2011). The row names and column names of this matrix corresponds to the genotype identifiers for Parent 1 and Parent 2.

[Package lme4GS version 1.0 Index]