ranef_uvcov {lme4GS}	R Documentation

Extract the conditional means of the random effects

Description

A function to extract the conditional means of the random effects from a fitted model object.

Usage

  ranef_uvcov(object)

Arguments

object
is an object returned by lmer_uvcov.

Details

The function ranef extract the conditional means for the liner mixed effects model:
y=Xβ + Z1* u1*+...+Zq* uq*+e,
where Zj*=Zj Lj, with Lj from Cholesky factorization for Kj. Alternatively, Zj*=ZjΓjΛ1/2, with Γj and Λj the matrix of eigen-vectors and eigen-values obtained from the eigen-value decomposition for Kj. So, the conditional means of the random effects in the linear mixed effects model:
y=Xβ + Z1 u1 + ... + Zq uq + e,
are obtained as follows: ûj=Ljûj* if the Cholesky factorization is used or ûj=ΓjΛj1/2ûj* if the the eigen-value decomposition is used.

Value

A list of data frames, one for each grouping factor.

Author(s)

Paulino Perez-Rodriguez

Examples

## Not run: 
  library(BGLR)
  library(lme4GS)

  #Example 1, wheat 
  data(wheat)
  X<-wheat.X
  Z<-scale(X,center=TRUE,scale=TRUE)
  G<-tcrossprod(Z)/ncol(Z)
  A<-wheat.A
  rownames(G)<-colnames(G)<-rownames(A)
  y<-wheat.Y[,1]

  #id a vector with ids
  random<-list(mrk=list(K=G,id=rownames(G)),
               ped=list(K=A,id=rownames(A)))

  out<-lmer_uvcov(y,fixed="1",random=random)

  out2<-ranef_uvcov(out)

  str(out2)

## End(Not run)

[Package lme4GS version 1.0 Index]