theta_optim {lme4GS}	R Documentation

Selection of bandwidth parameter of Gaussian kernel.

Description

Obtain the optimal value of the bandwidth parameter for the Gaussian kernel by cross-validation.

Usage

  theta_optim<-function(y,X)

Arguments

y
(numeric, n) the data-vector (NAs not allowed).

X
A matrix, n p that specifies the markers.

Value

A list that contains

theta
A vector that specifies the different values of theta.

theta.max
The optimal value of theta that maximizes the log-likelihood.

LL
A vector that specifies the values of the likelihood with its corresponding values of theta.

max.LL
The maximum value of the log-likelihood.

K
Tha Gassuian Kernel compute with the optimal value of theta.

Author(s)

Paulino Perez-Rodriguez

Examples

## Not run: 
  library(BGLR)
  library(lme4GS)

#Example wheat

data(wheat)
X <- wheat.X 
A <- wheat.A 
y <- wheat.Y[,1]

rownames(X) <- rownames(A)

out_res <- theta_optim(y,X)

## End(Not run)

[Package lme4GS version 1.0 Index]